// src/receipts/generated/items.ts
export enum ItemCategory {
  battery,
  bottle,
  bottledSolution,
  component,
  medicine,
  naturalOre,
  naturalPlant,
  orePowder,
  oreRefined,
  other,
  part,
  plantPowder,
  plantSeed,
  solution,
}